BACKLOG-21996 Content Templates PoC
These changes modify the way that templates are resolved to use an applyOn property to determine if a content template is available.

Note: default template resolution is currently not working, and other issues might occur. Also no automated tests exist, but they will be required to make sure there are no regressions or issues.

Currently to test, a special branch of the npm-templates-web-blue is available that contains content templates.