BACKLOG-23081: Build in production mode (#196)
Build NPM Modules Engine with the Webpack mode set to "production".
Add granularity for the scripts defined in package.json to be able to build only the client or the server side. Also, provide development versions of those scripts (by default, the production mode is used).
Bind a yarn clean to the mvn clean phase.